% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pseudo.R
\name{match_pseudo}
\alias{match_pseudo}
\title{Predict pseudotemporal information}
\usage{
match_pseudo(input_matrix, ref_matrix = NULL, standardize, species = "mouse")
}
\arguments{
\item{x}{cell data set (count matrix)}
}
\value{
predicted numeric pseudotime and branch
}
\description{
Predict pseudotemporal information
}
