% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pseudo.R
\name{xs}
\alias{xs}
\title{Predict pseudotemporal information}
\usage{
xs(x)
}
\arguments{
\item{x}{cell data set (count matrix)}
}
\value{
predicted numeric pseudotime and branch
}
\description{
Predict pseudotemporal information
}
